% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\alias{counties_inset}
\title{Outline of US counties}
\format{both objects are data frames consisting of variables of length ~50k:
\itemize{
  \item long: geographic longitude
  \item lat: geographic latitude
  \item order: order of locations
  \item hole: logical vector
  \item piece: positive integer value, one for each separate area/island of a county
  \item id: integer value uniquely identifying each county
  \item group: character value - composite of id and piece, uniquely identifying each part of each county
  \item STATEFP: FIPS state  code
  \item COUNTYFP: FIPS county  code
  \item COUNTYNS: GNIS county  code
  \item AFFGEOID: American FactFinder county code
  \item GEOID: GEOID county code
  \item NAME: name of county
  \item ALAND: land area of the county in square meters
  \item AWATER: water area of the county in square meters
  \item STATE: name of the state
  \item STUSPS: US state code
  \item REGION: integer of US region
  \item DIVISION: integer of US division
}}
\usage{
counties

counties_inset
}
\description{
Data from shapefiles provided by the Census Bureau’s MAF/TIGER geographic database at
\url{https://www.census.gov/geo/maps-data/data/cbf/cbf_state.html}. County outlines are based
on the 1:20m resolution level. Divisions and regions are based on TIGER/Line Shapefiles for states and equivalents in the 2016 release.
}
\examples{
counties \%>\% ggplot(aes(x = long, y = lat)) +
  geom_path(aes(group = group))

counties_inset \%>\% ggplot(aes(x = long, y = lat)) +
  geom_path(aes(group = group))
}
\keyword{datasets}

