% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crimes}
\alias{crimes}
\title{Crimes in the US since 1960}
\format{data frame of seven variables and about 23k rows:
\itemize{
  \item State: name of the state
  \item Abb: two-letter state abbreviation
  \item Year: year of the record
  \item Population: state population
  \item Type: type of crime committed
  \item Type2: type of crime committed distinguishing violent and property crimes
  \item Number: number of crimes
}}
\usage{
crimes
}
\description{
where does this dataset come from?
}
\examples{
library(magrittr)
crimes \%>\% dplyr::filter(Type=="Murder") \%>\%
  ggplot(aes(x = Year, y = Number/Population)) +
    geom_line(aes(group = State))
}
\keyword{datasets}
