% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{map_unif}
\alias{map_unif}
\title{Find a uniform sample of points for one region of a map}
\usage{
map_unif(map, n)
}
\description{
place uniform point sample into the region `region` of a map
}
\examples{
data(inset01)
df <- inset01 \%>\% filter(DIVISION == "9") \%>\% map_unif(1000)
df <- inset01 \%>\% filter(NAME == "Iowa") \%>\% map_unif(10)
df <- inset01 \%>\% map_unif(5000)
inset01 \%>\% ggplot(aes(x = long, y = lat)) +
  geom_path(aes(group = group)) +
  geom_point(data = df, colour = "red")

data(crimes)
population <- crimes \%>\% filter(Year == max(Year))
popmap <- left_join(inset, population[,c("Abb", "Population")], by=c("STUSPS"="Abb"))
poplist <- popmap \%>\% tidyr::nest(-NAME)
poplist$sample <- poplist$data \%>\% purrr::map(.f = function(d) d \%>\% map_unif(50))
poplist$sample <- poplist$data \%>\% purrr::map(.f = function(d) d \%>\% map_unif(round(d$Population[1]/50000)))
df <- poplist \%>\% select(-data) \%>\% tidyr::unnest()
inset \%>\% ggplot(aes(x = long, y = lat)) +
  geom_path(aes(group = group), size = 0.25) +
  geom_point(data = df, colour = "red", size = .5) +
  ggthemes::theme_map()
}

