% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_shape}
\alias{process_shape}
\title{read shape file}
\usage{
process_shape(path)
}
\arguments{
\item{path}{path to shapefile}
}
\description{
read shape file
}
\examples{
states <- process_shape("data-raw/tl_2016_us_state/tl_2016_us_state.shp")
states <- states \%>\% mutate(
long = replace(long, long > 100, long[long > 100]-360)
)
}
