% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a polygon shape by straightening out jittery lines}
\usage{
simplify(map, tol = 0.01)
}
\arguments{
\item{map}{map object as generated by functions such as `process_shape`, `ggplot2::map_data`, or `ggplot2::fortify`}

\item{tol}{tolerance to use for simplifying the map. Larger tolerances will lead to smaller maps at lower resolutions.}
}
\description{
Simplify a polygon shape by straightening out jittery lines
}
\examples{
states05 <- simplify(states, 0.001)
states05 \%>\% ggplot(aes(x = long, y = lat)) + geom_path(aes(group = group))
states01 <- simplify(states, 0.01)
states01 \%>\% ggplot(aes(x = long, y = lat)) + geom_path(aes(group = group))
}

