% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_rec}
\alias{simplify_rec}
\title{Douglas Pecker algorithm for line thinning}
\usage{
simplify_rec(x, y, tol = 0.01)
}
\arguments{
\item{x}{vector of longitudes}

\item{y}{vector of latitudes}

\item{tol}{tolerance}
}
\description{
Implementation of the Douglas-Peucker algorithm for line thinning
http://en.wikipedia.org/wiki/Ramer-Douglas-Peucker_algorithm
}

