% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{stat_polygon_jitter}
\alias{stat_polygon_jitter}
\title{ggplot2 stat for creating a sample of jittered points in a polygon}
\usage{
stat_polygon_jitter(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
ggplot2 stat for creating a sample of jittered points in a polygon
}

