% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{StatPolygon_jitter}
\alias{StatPolygon_jitter}
\title{ggplot2 stat for creating a sample of jittered points in a polygon}
\format{An object of class \code{StatPolygon_jitter} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatPolygon_jitter
}
\description{
ggplot2 stat for creating a sample of jittered points in a polygon
}
\examples{
data(division)
data(crimes)
library(dplyr)
crime.map <- left_join(division, crimes, by=c("NAME"="State"))
ggplot(data = crime.map, aes(x = long, y = lat)) +
  geom_polygon(aes(group = group), fill="grey90", colour="white", size = 0.5) +
  stat_polygon_jitter(aes(long = long, lat = lat, group = STATEFP,
                          mapgroup = group, n = Population/200000)) +
  ggthemes::theme_map()


}
\keyword{datasets}

