% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf.R
\name{unnest_sf}
\alias{unnest_sf}
\title{Convert multipolygon simple feature from sf format}
\usage{
unnest_sf(data, ...)
}
\arguments{
\item{data}{data set}

\item{...}{one or more columns to extract (one is tested, you're on your own with multiple)}
}
\description{
Works for multipolygon simple features (see package `sf`) only.
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
gnc <- sf_unnest(nc, geometry)
gnc \%>\% ggplot(aes(x = long, y = lat, group = group)) + geom_polygon()
}
