% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sf.R
\name{unnest_sf_}
\alias{unnest_sf_}
\title{Convert multipolygon simple feature from sf format}
\usage{
unnest_sf_(data, unnest_cols)
}
\arguments{
\item{data}{data set}

\item{unnest_cols}{one or more columns to extract (one is tested, you're on your own with multiple)}
}
\description{
Works for multipolygon simple features (see package `sf`) only.
}
