% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{scale}
\alias{scale}
\title{Scale regions in a map}
\usage{
scale(map, variable, regions, scale = 1, set_to = NULL)
}
\arguments{
\item{map}{map object as generated by `process_shape`  or `ggplot2::map_data`}

\item{variable}{name of the variable to filter on}

\item{regions}{name(s) of the levels in variable to use for the filter}

\item{scale}{numeric vector of length 2 used to scale the map region in longitude and latitude. Scale factors should be positive (negative values flip region).}

\item{set_to}{numeric vector of length 2. Set center of the region (defined by range in lat and long) to this longitude and latitude}
}
\description{
Scale regions in a map
}
\examples{
states01 <- states01 \%>\% group_by(group) \%>\% mutate(bbox = diff(range(long))*diff(range(lat)))
states01 <- states01 \%>\% filter(bbox > 0.15)
states01 \%>\% ggplot(aes(x = long, y = lat)) + geom_path(aes(group = group))
inset01 <- shift(states01, "NAME", "Hawaii", shift_by = c(51, 5.5))
inset01 <- scale(inset01, "NAME", "Alaska", scale=0.3, set_to=c(-120, 27))
inset01  \%>\% ggplot(aes(long, lat)) + geom_path(aes(group=group))
inset01 <- inset01  \%>\%
  filter(lat > 20)
inset01 \%>\%
  ggplot(aes(long, lat)) + geom_path(aes(group=group))
}

