% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify a polygon shape by straightening out jittery lines}
\usage{
simplify(map, tol = 0.01)
}
\description{
Simplify a polygon shape by straightening out jittery lines
}
\examples{
states05 <- simplify(states, 0.001)
states05 \%>\% ggplot(aes(x = long, y = lat)) + geom_path(aes(group = group))
states01 <- simplify(states, 0.01)
states01 \%>\% ggplot(aes(x = long, y = lat)) + geom_path(aes(group = group))
}

